# packages/schemas/api/v1/openapi.json
{
  "components": {
    "schemas": {
      "CredentialRotateRequest": {
        "additionalProperties": false,
        "properties": {
          "identity_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Identity Id",
            "type": "string"
          },
          "refresh_token": {
            "maxLength": 256,
            "minLength": 20,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "identity_id",
          "refresh_token"
        ],
        "title": "CredentialRotateRequest",
        "type": "object"
      },
      "GuestCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "device_public_key": {
            "maxLength": 128,
            "minLength": 32,
            "title": "Device Public Key",
            "type": "string"
          },
          "turnstile_proof": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Turnstile Proof",
            "type": "string"
          }
        },
        "required": [
          "turnstile_proof",
          "device_public_key"
        ],
        "title": "GuestCreateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrivacyOperation": {
        "enum": [
          "export",
          "correction",
          "deletion",
          "restriction",
          "legal_hold"
        ],
        "title": "PrivacyOperation",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LenGeas Identity and Tenancy API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounts/correction": {
      "post": {
        "operationId": "rights_api_v1_accounts_correction_post",
        "parameters": [
          {
            "in": "query",
            "name": "op",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PrivacyOperation",
              "default": "correction"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Rights Api V1 Accounts Correction Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rights"
      }
    },
    "/api/v1/accounts/deletion": {
      "post": {
        "operationId": "delete_account_api_v1_accounts_deletion_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Delete Account Api V1 Accounts Deletion Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Account"
      }
    },
    "/api/v1/accounts/export": {
      "post": {
        "operationId": "rights_api_v1_accounts_export_post",
        "parameters": [
          {
            "in": "query",
            "name": "op",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PrivacyOperation",
              "default": "export"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Rights Api V1 Accounts Export Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rights"
      }
    },
    "/api/v1/accounts/legal-hold": {
      "post": {
        "operationId": "rights_api_v1_accounts_legal_hold_post",
        "parameters": [
          {
            "in": "query",
            "name": "op",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PrivacyOperation",
              "default": "legal_hold"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Rights Api V1 Accounts Legal Hold Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rights"
      }
    },
    "/api/v1/accounts/restriction": {
      "post": {
        "operationId": "rights_api_v1_accounts_restriction_post",
        "parameters": [
          {
            "in": "query",
            "name": "op",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PrivacyOperation",
              "default": "restriction"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Rights Api V1 Accounts Restriction Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rights"
      }
    },
    "/api/v1/auth/credentials/rotate": {
      "post": {
        "operationId": "rotate_credential_api_v1_auth_credentials_rotate_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRotateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Rotate Credential Api V1 Auth Credentials Rotate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rotate Credential"
      }
    },
    "/api/v1/auth/current": {
      "get": {
        "operationId": "current_account_api_v1_auth_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Current Account Api V1 Auth Current Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current Account"
      }
    },
    "/api/v1/auth/guests": {
      "post": {
        "operationId": "create_guest_api_v1_auth_guests_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Guest Api V1 Auth Guests Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Guest"
      }
    },
...
    "/api/v1/privacy/requests/{request_id}": {
      "get": {
        "operationId": "privacy_status_api_v1_privacy_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Privacy Status Api V1 Privacy Requests  Request Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Privacy Status"
      }
    },
...
  }
}